+++ partial-recall-extensions.el
@@ -4,7 +4,8 @@
 
 ;;; Commentary:
 
-;; Extension to integrate `partial-recall' in external libraries.
+;; Functionality to integrate `partial-recall' in external libraries
+;; or make `partial-recall' accessible from external code.
 
 ;;; Code:
 
@@ -26,6 +27,39 @@
                                        :predicate #'partial-recall--memory-buffer-p
                                        :as #'buffer-name)))
   "Buffers that are recalled from the current tab.")
+
+;;;###autoload
+(defun partial-recall-buffer-specs (&optional buffer)
+  "Get the BUFFER's specs.
+
+The specs are a plist of the attributes `:meaningful', `:real'
+and `:implanted'."
+  (interactive)
+
+  (let* ((buffer (or buffer (current-buffer)))
+         (specs (list :meaningful (partial-recall--meaningful-buffer-p buffer)
+                      :real (not (null (partial-recall--memory-buffer-p buffer)))
+                      :implanted (buffer-local-value 'partial-recall--implanted buffer))))
+
+    (if (called-interactively-p 'any)
+        (partial-recall--message "Buffer '%s' has specs '%s'" buffer specs)
+      specs)))
+
+(defun partial-recall-memory-specs (&optional memory)
+  "Get the MEMORY's specs.
+
+The specs are a plist of attributes `:size' and `:capacity' and
+`:original-capacity'."
+  (interactive)
+
+  (let* ((memory (or memory (partial-recall--reality)))
+         (specs (list :size (ring-length (partial-recall--memory-ring memory))
+                      :capacity (ring-size (partial-recall--memory-ring memory))
+                      :original-capacity (partial-recall--memory-orig-size memory))))
+
+    (if (called-interactively-p 'any)
+        (partial-recall--message "Memory '%s' has specs '%s'" memory specs)
+      specs)))
 
 (provide 'partial-recall-extensions)